T: add types integer, real kind =8

+++ testdata/tqlrat.f.go
@@ -2,24 +2,24 @@ package main
 
 func tqlrat() {
 	{
-		var b real(kind=8)
-		var ubound.2 integer(kind=8)
+		var b float64
+		var ubound.2 int
 		ubound.2 = *n
-		var size.3 integer(kind=8)
+		var size.3 int
 		size.3 = ubound.2
 		size.3 = f4goUndefinedmax_expr
-		var tempVarF4GO_65 integer(kind=8)
+		var tempVarF4GO_65 int
 		tempVarF4GO_65 = size.3 + -1
 		var tempVarF4GO_67 bitsizetype
 		tempVarF4GO_67 = size.3 * 64
 		var tempVarF4GO_69 sizetype
 		tempVarF4GO_69 = size.3 * 8
-		var ubound.0 integer(kind=8)
+		var ubound.0 int
 		ubound.0 = *n
-		var size.1 integer(kind=8)
+		var size.1 int
 		size.1 = ubound.0
 		size.1 = f4goUndefinedmax_expr
-		var tempVarF4GO_76 integer(kind=8)
+		var tempVarF4GO_76 int
 		tempVarF4GO_76 = size.1 + -1
 		var tempVarF4GO_78 bitsizetype
 		tempVarF4GO_78 = size.1 * 64
@@ -32,9 +32,9 @@ func tqlrat() {
 			var i int
 			i = 2
 		}
-		var f real(kind=8)
+		var f float64
 		f = 0.0
-		var t real(kind=8)
+		var t float64
 		t = 0.0
 		(*e2)[*n+-1] = 0.0
 		{
